move ingress processing logic to separate structure
